feat(core/cell): - make simular cell constructor -update cell's  and players's coordinates storage

+++ src/core/render/camera.hpp
@@ -28,7 +28,7 @@ class Camera
 {
 public:
 
-    glm::vec3 Position;
+    glm::dvec3 Position;
     glm::vec3 Front;
     glm::vec3 Up;
     glm::vec3 Right;
@@ -42,7 +42,7 @@ public:
     float Zoom;
 
 
-    Camera(glm::vec3 position = glm::vec3(0.0f, 0.0f, 0.0f), glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f), float yaw = YAW, float pitch = PITCH);
+    Camera(glm::dvec3 position = glm::dvec3(0.0f, 0.0f, 0.0f), glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f), float yaw = YAW, float pitch = PITCH);
 
     Camera(float posX, float posY, float posZ, float upX, float upY, float upZ, float yaw, float pitch);
